test: update E2E scenarios for detecting ANRs
