+++ .buildkite/pipeline.yml
@@ -299,6 +299,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
 
@@ -317,6 +318,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
     soft_fail:
@@ -337,6 +339,7 @@ steps:
         - "--username=$BROWSER_STACK_USERNAME"
         - "--access-key=$BROWSER_STACK_ACCESS_KEY"
         - "--fail-fast"
+        - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
     #
@@ -360,6 +363,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
     #
@@ -383,6 +387,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
     soft_fail:
@@ -403,6 +408,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
     soft_fail:
@@ -423,6 +429,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
 
@@ -441,6 +448,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
     soft_fail:
@@ -461,6 +469,7 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
+          - "features/full"
     concurrency: 10
     concurrency_group: 'browserstack-app'
     soft_fail:
@@ -481,6 +490,6 @@ steps:
           - "--username=$BROWSER_STACK_USERNAME"
           - "--access-key=$BROWSER_STACK_ACCESS_KEY"
           - "--fail-fast"
-          - "features/unhandled_exception.feature"
+          - "features/minimal"
     concurrency: 10
     concurrency_group: 'browserstack-app'
