+++ features/full/bugsnag_init.feature
