+++ features/full/strict_mode.feature
