+++ features/full/trimmed_stacktrace.feature
